requirements.txt
openai==1.23.6
python-dotenv==1.0.1
